Should I use redis for caching?
