+++ client/src/components/preview.tsx
@@ -1,4 +1,5 @@
 import React, { useEffect, useRef } from 'react';
+import { iframeHTML } from 'utils/template';
 // don't use css modules on this file because it's using a special class name to refer the
 // react-resizable element
 import './preview.sass';
@@ -11,46 +12,12 @@ interface PreviewProps {
 const Preview: React.FC<PreviewProps> = ({ code, err }) => {
   const iframeRef = useRef<HTMLIFrameElement>(null);
 
-  const html = `
-    <html>
-      <head>
-        <style>
-          html {background-color: white;}
-          body {font-family: Arial, Helvetica, sans-serif;}
-        </style>
-      </head>
-      <body>
-        <div id="root"></div>
-        <script>
-          const handleError = (err) => {
-            const root = document.getElementById('root');
-            root.innerHTML = '<div style="color: red;"><h4>Runtime Error</h4>' + err + '</div>'
-            console.error(err);
-          }
-
-          window.addEventListener('error', (event) => {
-            event.preventDefault();
-            handleError(event.error);
-          });
-
-          window.addEventListener('message', (event) => {
-            try {
-              eval(event.data);
-            } catch(err) {
-              handleError(err);
-            } 
-          }, false)
-        </script>
-      </body>
-    </html>
-  `;
-
   useEffect(() => {
-    if (!iframeRef.current) {
-      return;
-    }
+    if (!iframeRef.current) return;
 
-    iframeRef.current.srcdoc = html;
+    // If for some crazy reason the user writes down `document.body.innertHTML = '';`
+    // this will recreate the default ui for an OK behaviour
+    iframeRef.current.srcdoc = iframeHTML;
 
     const timer = window.setTimeout(() => {
       iframeRef.current!.contentWindow?.postMessage(code, '*');
@@ -66,7 +33,7 @@ const Preview: React.FC<PreviewProps> = ({ code, err }) => {
       <iframe
         ref={iframeRef}
         sandbox="allow-scripts"
-        srcDoc={html}
+        srcDoc={iframeHTML}
         title="preview"
       />
       {err && <div className="preview-err">{err}</div>}
